REFACTOR: Remoção de Repository e correções crud
Removida a camada repository devido ao número de requisições que será feita para a aplicação.
Correções no CRUD, primeiro projeto em php 8 e laravel 9.

+++ database/migrations/2023_03_14_052630_create_produtos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('produtos', function (Blueprint $table) {
             $table->id();
             $table->string('nome', 60);
-            $table->integer('valor');
+            $table->double('valor');
             $table->unsignedBigInteger('loja_id');
             $table->boolean('ativo');
             $table->timestamps();
